requirements.txt
tensorflow==2.10.1
Pillow==9.1.1
numpy==1.23.3
